refactor code to separate steps from setup
